libs/miroflow-tools/pyproject.toml
[project]
name = "miroflow-tools"
version = "0.1.0"
description = "Tool management and MCP server utilities for MiroFlow"
readme = "README.md"
authors = [
    { name = "MiroMind Team", email = "service@miromind.ai" }
]
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0.0",
    "fastmcp>=0.1.0",
    "playwright>=1.40.0",
    "requests>=2.32.0",
    "e2b-code-interpreter==1.2.1",
    "wikipedia",
    "mutagen",
    "markitdown-mcp>=0.0.1a3",
    "google-genai",
    "aiohttp",
    "redis",
    "Pillow>=10.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/miroflow_tools"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
    "pytest-html>=4.1.1",
    "pytest-xdist>=3.7.0",
    "pytest-mock>=3.10.0",
    "pytest-timeout>=2.1.0",
    "inline-snapshot>=0.23.2",
]

[tool.pytest.ini_options]
minversion = "8.3.5"
testpaths = ["src/test"]
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-rA",
    "--show-capture=stderr",
    "-n=auto",
    "--html=report.html",
    "--self-contained-html",
    "--cov=miroflow_tools",
    "--cov-report=html",
    "--strict-markers",
    "-v",
]
markers = [
    "integration: marks tests as integration tests (may be slow)",
    "unit: marks tests as unit tests",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "requires_api_key: marks tests that require real API credentials",
] 